% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractInteractions.R
\name{coef.EMglmnet}
\alias{coef.EMglmnet}
\title{Extract Model Coefficients From a Fitted Mixture of Experts Model}
\usage{
\method{coef}{EMglmnet}(object, ...)
}
\arguments{
\item{object}{[\code{\link{EMglmnet}}]\cr
An object of class \code{EMglmnet} of which to extract the model coefficients.}

\item{...}{[\code{ANY}]\cr
Arguments to \code{\link[glmnet]{coef.glmnet}}.}
}
\value{
A list with model coefficients belonging to the gating and expert models.
}
\description{
Extract model coefficients from a fitted mixture of experts model.
}

