% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMlnet.R
\name{EMlnet}
\alias{EMlnet}
\title{Penalized Mixtures of Experts}
\usage{
EMlnet(y, X, colsGating = 1:ncol(X), colsExperts = 1:ncol(X), J = 2,
  lambda, alpha = 1, standardize = FALSE, interceptGating = TRUE,
  interceptExperts = TRUE, offsetGating = NULL, offsetExperts = NULL,
  iter.max = 100, tol = 1e-06, init.weights = NULL)
}
\arguments{
\item{y}{A \code{factor} specifying the class membership for each observation.}

\item{X}{A \code{matrix} or \code{data.frame} or \code{Matrix} containing
the explanatory variables (must not contain an intercept column).}

\item{colsGating}{Names or indices of columns in \code{X} to be used for the gating model. Default is all columns.}

\item{colsExperts}{Names or indices of columns in \code{X} to be used for the expert models. Default is all columns.}

\item{J}{The number of experts / mixture components. Defaults to 2.}

\item{lambda}{Penalty parameter. Can be a scalar or a vector of length \code{1+J} with different components for the gating and
the \code{J} expert models. All components must be >= 0.}

\item{alpha}{Mixing parameter for the elastic net penalty. Can be a scalar or a vector of length \code{1+J} with different
components for the gating and the \code{J} expert models. All components must be in [0,1].
Defaults to 1.}

\item{standardize}{Logical. Should the columns of \code{X} be standardized prior to fitting the model?}

\item{interceptGating}{Logical. Does the gating model include an intercept?
Defaults to \code{TRUE}.}

\item{interceptExperts}{Logical. Does the expert model include an intercept?
Defaults to \code{TRUE}.}

\item{offsetGating}{Offset term for the gating model.}

\item{offsetExperts}{Offset term for the expert model.}

\item{iter.max}{Maximum number of EM iterations.}

\item{tol}{Small positive value to detect convergence of the EM algorithm.}

\item{init.weights}{An \code{nrow(X)} times \code{J} matrix containing initial group membership values for the EM algorithm.
The row sums must be 1. Defaults to \code{NULL}, which means that initial weights will be obtained by random segmentation.}
}
\value{
An object of class \code{EMlnet}.
 A \code{list} with entries:
  \item{gating}{The gating model. An object of class \code{glmnet}.}
  \item{experts}{A \code{list} of \code{J} expert models. These are objects of class \code{glmnet}}
  \item{weights}{The final group membership values.}
  \item{J,K, colsGating, colsExperts, lambda, offsetGating, offsetExperts}{See arguments.}
  \item{lev1}{Class labels present in the data.}
  \item{lev}{Class labels.}
}
\description{
Fits a penalized mixture of experts model using the EM algorithm.
In the EM steps penalized linear models are fitted using function \code{\link[glmnet]{glmnet}} from \pkg{glmnet} is used.
}

