% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractInteractions.R
\name{getInteractions}
\alias{getInteractions}
\title{Extract Interactions From a Fitted Mixture of Experts Model}
\usage{
getInteractions(object, ...)
}
\arguments{
\item{object}{[\code{\link{EMglmnet}}]\cr
An object of class \code{EMglmnet} of which to extract the model coefficients.}

\item{...}{[\code{ANY}]\cr
Arguments to \code{\link[glmnet]{coef.glmnet}}.}
}
\value{
A \code{data.frame} of active interactions.
}
\description{
Extract interactions from a fitted mixture of experts model.
}

