% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractInteractions.R
\name{plotPath}
\alias{plotPath}
\title{Path Plots for a Fitted Mixture of Experts Model}
\usage{
plotPath(x, ...)
}
\arguments{
\item{x}{[\code{\link{EMglmnet}}]\cr
An object of class \code{EMglmnet} of which to extract the model coefficients.}

\item{...}{[\code{ANY}]\cr
Arguments to \code{\link[glmnet]{plot.glmnet}}.}
}
\description{
Path plots for a fitted mixture of experts model.
}

