% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMglmnet.R
\name{predict.EMglmnet}
\alias{predict.EMglmnet}
\title{Predict New Observations by a Trained Logistic Mixture of Experts Model}
\usage{
\method{predict}{EMglmnet}(object, newdata, ...)
}
\arguments{
\item{object}{An object of class \code{EMglmnet}.}

\item{newdata}{A \code{data.frame} or \code{matrix} with data for which to make predictions.}

\item{...}{Further arguments. Currently unused.}
}
\value{
A \code{list} with components:
\item{class}{A \code{factor} with predicted class levels.}
\item{posterior}{A \code{matrix} with predicted posterior probabilities.}
\item{gating}{The probabilities predicted by the fating model.}
\item{experts}{The class posterior probabilities predicted by individual experts.}
}
\description{
Predict new observations by a trained logistic mixture of experts model.
}

